create DataDownloadSpec
