+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/test/metadata/DataSpec.scala
@@ -1,61 +1,22 @@
 package org.broadinstitute.dsde.firecloud.test.metadata
 
-import org.broadinstitute.dsde.workbench.config.{Config, UserPool}
+import org.broadinstitute.dsde.workbench.config.UserPool
 import java.io.{File, PrintWriter}
 import java.util.UUID
 
 import org.broadinstitute.dsde.firecloud.fixture.{TestData, UserFixtures}
 import org.broadinstitute.dsde.firecloud.page.workspaces.WorkspaceDataPage
 import org.broadinstitute.dsde.firecloud.page.workspaces.methodconfigs.WorkspaceMethodConfigDetailsPage
-import org.broadinstitute.dsde.firecloud.page.workspaces.summary.WorkspaceSummaryPage
 import org.broadinstitute.dsde.workbench.auth.AuthToken
 import org.broadinstitute.dsde.workbench.fixture._
-import org.broadinstitute.dsde.workbench.service.{AclEntry, Orchestration, WorkspaceAccessLevel}
+import org.broadinstitute.dsde.workbench.service.{AclEntry, WorkspaceAccessLevel}
 import org.broadinstitute.dsde.workbench.service.test.{CleanUp, WebBrowserSpec, WebBrowserUtil}
-import org.openqa.selenium.WebDriver
 import org.scalatest.selenium.WebBrowser
 import org.scalatest.{FreeSpec, Matchers}
 
-import scala.io.Source
 
-class DataSpec extends FreeSpec with WebBrowserSpec
-  with UserFixtures with WorkspaceFixtures with BillingFixtures
+class DataSpec extends FreeSpec with WebBrowserSpec with UserFixtures with WorkspaceFixtures with BillingFixtures
   with Matchers with WebBrowser with WebBrowserUtil with CleanUp {
-
-  private def makeTempDownloadDirectory(): String = {
-    /*
-     * This might work some day if docker permissions get straightened out... or it might not be
-     * needed. For now, we instead `chmod 777` the directory in run-tests.sh.
-    new File("chrome").mkdirs()
-    val downloadPath = Files.createTempDirectory(Paths.get("chrome"), "downloads")
-    val permissions = Set(PosixFilePermission.OWNER_WRITE, PosixFilePermission.GROUP_WRITE, PosixFilePermission.OTHERS_WRITE)
-    Files.setPosixFilePermissions(downloadPath, permissions.asJava)
-    downloadPath.toString
-     */
-    val downloadPath = "chrome/downloads"
-    new File(downloadPath).mkdirs()
-    downloadPath
-  }
-
-  private val downloadPath = makeTempDownloadDirectory()
-//  val owner: Credentials = UserPool.chooseProjectOwner
-//  val reader: Credentials = UserPool.chooseStudent
-//  implicit lazy val authToken: AuthToken = owner.makeAuthToken()
-
-  "import a participants file" in withWebDriver { implicit driver =>
-    val owner = UserPool.chooseProjectOwner
-    implicit val authToken: AuthToken = owner.makeAuthToken()
-    withCleanBillingProject(owner) { billingProject =>
-      withWorkspace(billingProject, "TestSpec_FireCloud_import_participants_file_") { workspaceName =>
-        withSignIn(owner) { _ =>
-          val filename = "src/test/resources/participants.txt"
-          val workspaceDataTab = new WorkspaceDataPage(billingProject, workspaceName).open
-          workspaceDataTab.importFile(filename)
-          workspaceDataTab.getNumberOfParticipants shouldBe 1
-        }
-      }
-    }
-  }
 
   val methodConfigName: String = SimpleMethodConfig.configName + "_" + UUID.randomUUID().toString
 
@@ -364,84 +325,9 @@ class DataSpec extends FreeSpec with WebBrowserSpec
     }
   }
 
-  "Download should reflect visible columns" - {
-    "no workspace defaults or user preferences" in withWebDriver(downloadPath) { implicit driver =>
-      testMetadataDownload(
-        initialColumns = List("participant_id", "foo"),
-        expectedColumns = List("participant_id", "foo"))
-    }
 
-    "no workspace defaults, with user preferences" in withWebDriver(downloadPath) { implicit driver =>
-      testMetadataDownload(
-        initialColumns = List("participant_id", "foo"),
-        userHidden = Some("foo"),
-        expectedColumns = List("participant_id"))
-    }
-
-    "no workspace defaults, with user preferences, new columns" in withWebDriver(downloadPath) { implicit driver =>
-      testMetadataDownload(
-        initialColumns = List("participant_id", "foo"),
-        userHidden = Some("foo"),
-        importColumns = Some(List("participant_id", "foo", "bar")),
-        expectedColumns = List("participant_id", "bar"))
-    }
-
-    "with workspace defaults, no user preferences" in withWebDriver(downloadPath) { implicit driver =>
-      testMetadataDownload(
-        initialColumns = List("participant_id", "foo", "bar"),
-        defaultShown = Some(List("participant_id", "foo")),
-        defaultHidden = Some(List("bar")),
-        expectedColumns = List("participant_id", "foo"))
-    }
-
-    "with workspace defaults, no user preferences, new columns" in withWebDriver(downloadPath) { implicit driver =>
-      testMetadataDownload(
-        initialColumns = List("participant_id", "foo", "bar"),
-        defaultShown = Some(List("participant_id", "foo")),
-        defaultHidden = Some(List("bar")),
-        importColumns = Some(List("participant_id", "foo", "bar", "baz")),
-        expectedColumns = List("participant_id", "foo", "baz"))
-    }
-
-    "with workspace defaults, with user preferences" in withWebDriver(downloadPath) { implicit driver =>
-      testMetadataDownload(
-        initialColumns = List("participant_id", "foo", "bar"),
-        defaultShown = Some(List("participant_id", "foo")),
-        defaultHidden = Some(List("bar")),
-        userHidden = Some("foo"),
-        expectedColumns = List("participant_id"))
-    }
-
-    "with workspace defaults, with user preferences, new columns" in withWebDriver(downloadPath) { implicit driver =>
-      testMetadataDownload(
-        initialColumns = List("participant_id", "foo", "bar"),
-        defaultShown = Some(List("participant_id", "foo")),
-        defaultHidden = Some(List("bar")),
-        userHidden = Some("foo"),
-        importColumns = Some(List("participant_id", "foo", "bar", "baz")),
-        expectedColumns = List("participant_id", "baz"))
-    }
-
-    "keep ID column in download even if hidden in UI" in withWebDriver(downloadPath) { implicit driver =>
-      val user = UserPool.chooseAnyUser
-      implicit val authToken: AuthToken = user.makeAuthToken()
-      withCleanBillingProject(user) { billingProject =>
-        withWorkspace(billingProject, "DataSpec_download") { workspaceName =>
-          withSignIn(user) { _ =>
-            val dataTab = new WorkspaceDataPage(billingProject, workspaceName).open
-            val columns = List("participant_id", "foo")
-            dataTab.importFile(generateMetadataFile(columns))
-            dataTab.dataTable.hideColumn("participant_id")
-            val metadataFile = dataTab.downloadMetadata(Option(downloadPath)).get
-            readHeadersFromTSV(metadataFile) shouldEqual columnsToFileHeaders(columns)
-          }
-        }
-      }
-    }
-  }
-
-//  This test is just to make sure functionality in this context works
-//  BUT we should really also write some tests for this specific component (seperate of this context)
+  //  This test is just to make sure functionality in this context works
+  //  BUT we should really also write some tests for this specific component (seperate of this context)
   "Column reordering should be reflected" in withWebDriver {implicit driver =>
     val user = UserPool.chooseProjectOwner
     implicit val authToken: AuthToken = user.makeAuthToken()
@@ -458,69 +344,4 @@ class DataSpec extends FreeSpec with WebBrowserSpec
     }
   }
 
-  private def testMetadataDownload(initialColumns: List[String],
-                                   defaultShown: Option[List[String]] = None,
-                                   defaultHidden: Option[List[String]] = None,
-                                   userHidden: Option[String] = None,
-                                   importColumns: Option[List[String]] = None,
-                                   expectedColumns: List[String])
-                                  (implicit webDriver: WebDriver): Unit = {
-    val owner = UserPool.chooseProjectOwner
-    val writer = UserPool.chooseStudent
-    implicit val authToken: AuthToken = owner.makeAuthToken()
-
-    withCleanBillingProject(owner) { billingProject =>
-      withWorkspace(billingProject, "DataSpec_download", aclEntries = List(AclEntry(writer.email, WorkspaceAccessLevel.Writer))) { workspaceName =>
-        withSignIn(owner) { _ =>
-          (defaultShown, defaultHidden) match {
-            case (None, None) => Unit
-            case (_, _) => setColumnDefaults(billingProject, workspaceName, defaultShown.getOrElse(List()), defaultHidden.getOrElse(List()))
-          }
-
-          val dataTab = new WorkspaceDataPage(billingProject, workspaceName).open
-          dataTab.importFile(generateMetadataFile(initialColumns))
-          userHidden.foreach(dataTab.dataTable.hideColumn)
-        }
-
-        withSignIn(writer) { _ =>
-          val dataTab = new WorkspaceDataPage(billingProject, workspaceName).open
-          userHidden.foreach(dataTab.dataTable.hideColumn)
-
-          importColumns foreach { l => dataTab.importFile(generateMetadataFile(l)) }
-
-          dataTab.dataTable.readColumnHeaders shouldEqual expectedColumns
-          val metadataFile = dataTab.downloadMetadata(Option(downloadPath)).get
-          readHeadersFromTSV(metadataFile) shouldEqual columnsToFileHeaders(expectedColumns)
-        }
-
-        withSignIn(owner) { _ =>
-          val dataTab = new WorkspaceDataPage(billingProject, workspaceName).open
-          dataTab.dataTable.readColumnHeaders shouldEqual expectedColumns
-          val metadataFile = dataTab.downloadMetadata(Option(downloadPath)).get
-          readHeadersFromTSV(metadataFile) shouldEqual columnsToFileHeaders(expectedColumns)
-        }
-      }
-    }
-  }
-
-  def setColumnDefaults(billingProject: String, workspaceName: String, shown: List[String], hidden: List[String])
-                       (implicit authToken: AuthToken): Unit = {
-
-    def buildColumnDefaults: String = {
-      def join(columns: List[String]) = columns match {
-        case List() => ""
-        case l => "\"" + s"${columns.mkString("\", \"")}" + "\""
-      }
-      s"""{"participant": {"shown": [${join(shown)}], "hidden": [${join(hidden)}]}}"""
-    }
-
-    api.workspaces.setAttributes(billingProject, workspaceName, Map("workspace-column-defaults" -> buildColumnDefaults))
-  }
-  def readHeadersFromTSV(fileName: String): List[String] = {
-    Source.fromFile(fileName).getLines().next().split('\t').toList
-  }
-
-  def columnsToFileHeaders(columns: List[String]): List[String] = {
-    List("entity:" + columns.head) ++ columns.tail
-  }
 }
